@@ -12,13 +12,13 @@
   "main": "dist/Recompose.cjs.js",
   "module": "dist/Recompose.esm.js",
   "dependencies": {
-    "@babel/runtime": "^7.0.0",
-    "change-emitter": "^0.1.2",
-    "hoist-non-react-statics": "^2.3.1",
-    "react-lifecycles-compat": "^3.0.2",
-    "symbol-observable": "^1.0.4"
+    "@babel/runtime": "^7.24.5",
+    "change-emitter": "^0.1.6",
+    "hoist-non-react-statics": "^3.3.2",
+    "react-lifecycles-compat": "^3.0.4",
+    "symbol-observable": "^4.0.0"
   },
   "peerDependencies": {
-    "react": "^0.14.0 || ^15.0.0 || ^16.0.0"
+    "react": "^18.3.1"
   }
 }
