@@ -17,11 +17,11 @@
   "main": "cjs/RecomposeRelay.js",
   "module": "es/RecomposeRelay.js",
   "dependencies": {
-    "lodash": "^4.0.0"
+    "lodash": "^4.17.21"
   },
   "peerDependencies": {
-    "recompose": "^0.17.0 || ^0.18.0 || ^0.19.0 || ^0.20.0",
-    "react": "^0.14.0 || ^15.0.0",
-    "react-relay": "^0.6.0 || ^0.7.0 || ^0.8.0 || ^0.9.0"
+    "@beegru/recompose": "^0.30.1",
+    "react": "^18.3.1",
+    "react-relay": "^16.2.0"
   }
 }
